+++ tests/Feature/BreakpointTest.php
@@ -2,19 +2,11 @@
 
 namespace Spatie\ResponsiveImages\Tests\Feature;
 
-use Facades\Statamic\Imaging\GlideServer;
 use Illuminate\Http\UploadedFile;
-use Illuminate\Support\Collection;
-use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
-use Spatie\ResponsiveImages\AssetNotFoundException;
-use Spatie\ResponsiveImages\Responsive;
 use Spatie\ResponsiveImages\Tests\TestCase;
 use Spatie\ResponsiveImages\Breakpoint;
-use Statamic\Assets\AssetContainer;
 use Statamic\Facades\Stache;
-use Statamic\Fields\Value;
-use Statamic\Tags\Parameters;
 
 class BreakpointTest extends TestCase
 {
@@ -25,32 +17,13 @@ class BreakpointTest extends TestCase
     {
         parent::setUp();
 
-        config(['filesystems.disks.test' => [
-            'driver' => 'local',
-            'root' => __DIR__.'/tmp',
-            'url' => '/test',
-        ]]);
-
-        /** @var \Statamic\Assets\AssetContainer $assetContainer */
-        $assetContainer = (new AssetContainer)
-            ->handle('test_container')
-            ->disk('test')
-            ->save();
-
         Storage::disk('test')->delete('*');
 
         $file = new UploadedFile($this->getTestJpg(), 'test.jpg');
         $path = ltrim('/'.$file->getClientOriginalName(), '/');
-        $this->asset = $assetContainer->makeAsset($path)->upload($file);
+        $this->asset = $this->assetContainer->makeAsset($path)->upload($file);
 
         Stache::clear();
-    }
-
-    protected function tearDown(): void
-    {
-        parent::tearDown();
-
-        File::deleteDirectory(__DIR__ . '/tmp');
     }
 
     /** @test * */
@@ -60,7 +33,7 @@ class BreakpointTest extends TestCase
 
         $this->assertStringContainsString(
             '?w=100',
-            $responsive->buildImage(100)
+            $responsive->buildImageJob(100)->handle()
         );
     }
 
@@ -71,7 +44,7 @@ class BreakpointTest extends TestCase
 
         $this->assertStringContainsString(
             '?fm=webp&w=100',
-            $responsive->buildImage(100, ['fm' => 'webp'])
+            $responsive->buildImageJob(100, 'webp')->handle()
         );
     }
 
@@ -82,7 +55,7 @@ class BreakpointTest extends TestCase
 
         $this->assertStringContainsString(
             '?fm=webp&w=100&h=100',
-            $responsive->buildImage(100, ['fm' => 'webp'], 1.0)
+            $responsive->buildImageJob(100, 'webp', 1.0)->handle()
         );
     }